// apps/front-end/src/http/types/GetProductResponseDto.ts
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 */

export type GetProductResponseDto = {
  /**
   * @type object
   */
  product: {
    /**
     * @type string
     */
    id: string;
    /**
     * @type string
     */
    description: string;
    /**
     * @type string, uri
     */
    imageUrl: string;
    /**
     * @type string
     */
    imageBlurData: string;
    /**
     * @type number
     */
    price: number;
    /**
     * @type string
     */
    title: string;
    /**
     * @type string
     */
    organizationId: string;
    /**
     * @type string
     */
    createdAt?: string | null;
    /**
     * @type string
     */
    updatedAt?: string | null;
  };
  /**
   * @type string | undefined
   */
  message?: string;
};