// apps/front-end/src/http/types/UpdateProductBodyDto.ts
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 */

export type UpdateProductBodyDto = {
  /**
   * @type string | undefined
   */
  image?: string;
  /**
   * @type string | undefined
   */
  title?: string;
  /**
   * @type string | undefined
   */
  description?: string;
  /**
   * @type number | undefined
   */
  price?: number;
};